fix: address Hive PR review comments
Resolves review thread PRRT_kwDORy0Fc86A0sQB

Resolves review thread PRRT_kwDORy0Fc86A0sPo

Resolves review thread PRRT_kwDORy0Fc86A0sP1

Resolves review thread PRRT_kwDORy0Fc86A0sP4

Resolves review thread PRRT_kwDORy0Fc86A0sP6

Resolves review thread PRRT_kwDORy0Fc86A0sP9

Resolves review thread PRRT_kwDORy0Fc86A0oN_

Resolves review thread PRRT_kwDORy0Fc86A0oOB

Resolves review thread PRRT_kwDORy0Fc86A0oOC

Resolves review thread PRRT_kwDORy0Fc86A0nno

Resolves review thread PRRT_kwDORy0Fc86A0nnt

Resolves review thread PRRT_kwDORy0Fc86A0nnf

Resolves review thread PRRT_kwDORy0Fc86A0nnk
