+++ src/bin/syslog.rs
@@ -1,7 +1,4 @@
-#[path = "../main.rs"]
-mod hive_main;
-
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    hive_main::entry().await
+    hive_mcp::entry().await
 }
