+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod api;
 pub mod app;
+pub(crate) mod cli;
 pub mod config;
 pub mod mcp;
 pub mod observability;
@@ -10,6 +11,223 @@ pub mod syslog;
 pub(crate) mod db;
 pub(crate) mod docker_ingest;
 pub(crate) mod ingest;
+
+use anyhow::Result;
+use axum::Router;
+use rmcp::{transport::stdio, ServiceExt};
+use tracing::info;
+use tracing_subscriber::{fmt, EnvFilter};
+
+pub async fn entry() -> Result<()> {
+    let mode = Mode::parse(std::env::args().skip(1).collect())?;
+    if mode == Mode::Help {
+        print_usage();
+        return Ok(());
+    }
+    if mode == Mode::Version {
+        println!("Hive {}", env!("CARGO_PKG_VERSION"));
+        return Ok(());
+    }
+
+    fmt()
+        .with_env_filter(
+            EnvFilter::try_from_default_env()
+                .unwrap_or_else(|_| EnvFilter::new(mode.default_log_filter())),
+        )
+        .with_writer(std::io::stderr)
+        .with_target(true)
+        .init();
+
+    info!("Hive v{}", env!("CARGO_PKG_VERSION"));
+
+    match mode {
+        Mode::ServeMcp => serve_mcp().await,
+        Mode::StdioMcp => serve_stdio_mcp().await,
+        Mode::Cli(command) => run_cli(command).await,
+        Mode::Help => unreachable!("handled before logging initialization"),
+        Mode::Version => unreachable!("handled before logging initialization"),
+    }
+}
+
+async fn serve_stdio_mcp() -> Result<()> {
+    let runtime = runtime::RuntimeCore::load_query_only().await?;
+    let service = mcp::rmcp_server(runtime.mcp_state()).serve(stdio()).await?;
+    service.waiting().await?;
+    Ok(())
+}
+
+async fn run_cli(command: cli::CliCommand) -> Result<()> {
+    let runtime = runtime::RuntimeCore::load_query_only().await?;
+    cli::run(runtime.service(), command).await
+}
+
+async fn serve_mcp() -> Result<()> {
+    let runtime = runtime::RuntimeCore::load().await?;
+    info!(
+        syslog_bind = %runtime.config.syslog.bind_addr(),
+        mcp_bind = %runtime.config.mcp.bind_addr(),
+        db_path = %runtime.config.storage.db_path.display(),
+        retention_days = runtime.config.storage.retention_days,
+        max_db_size_mb = runtime.config.storage.max_db_size_mb,
+        recovery_db_size_mb = runtime.config.storage.recovery_db_size_mb,
+        min_free_disk_mb = runtime.config.storage.min_free_disk_mb,
+        recovery_free_disk_mb = runtime.config.storage.recovery_free_disk_mb,
+        cleanup_interval_secs = runtime.config.storage.cleanup_interval_secs,
+        pool_size = runtime.config.storage.pool_size,
+        wal_mode = runtime.config.storage.wal_mode,
+        mcp_auth_enabled = runtime.config.mcp.api_token.is_some(),
+        api_enabled = runtime.config.api.enabled,
+        docker_ingest_enabled = runtime.config.docker_ingest.enabled,
+        docker_ingest_hosts = runtime.config.docker_ingest.hosts.len(),
+        "Configuration loaded"
+    );
+
+    runtime.start_syslog().await?;
+    let _maintenance = runtime.spawn_maintenance_tasks();
+
+    let mut app: Router = mcp::router(runtime.mcp_state());
+    if runtime.config.api.enabled {
+        app = app.merge(api::router(api::ApiState {
+            service: runtime.service(),
+            config: runtime.config.api.clone(),
+            cors_port: runtime.config.mcp.port,
+            auth_policy: runtime.auth_policy().clone(),
+        })?);
+        info!("Non-MCP API mounted under /api");
+    }
+    app = app.merge(runtime.otlp_router());
+    info!("OTLP receiver mounted at /v1/logs (and /v1/metrics -> 200, /v1/traces -> 404)");
+    if runtime.config.mcp.api_token.is_none() && !runtime.config.mcp.host.starts_with("127.") {
+        tracing::warn!(
+            bind = %runtime.config.mcp.bind_addr(),
+            "OTLP /v1/logs is mounted WITHOUT authentication on a non-loopback bind. \
+             Anyone reachable on this address can write log records. \
+             Set HIVE_MCP_TOKEN or legacy SYSLOG_MCP_TOKEN to require Bearer auth."
+        );
+    }
+    app = app.layer(tower_http::trace::TraceLayer::new_for_http());
+
+    let mcp_bind = runtime.config.mcp.bind_addr();
+    let listener = tokio::net::TcpListener::bind(&mcp_bind).await?;
+    info!(bind = %mcp_bind, "MCP server listening");
+
+    // OTLP handler needs ConnectInfo<SocketAddr> for source_ip provenance.
+    axum::serve(
+        listener,
+        app.into_make_service_with_connect_info::<std::net::SocketAddr>(),
+    )
+    .with_graceful_shutdown(shutdown_signal())
+    .await?;
+
+    Ok(())
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+enum Mode {
+    ServeMcp,
+    StdioMcp,
+    Cli(cli::CliCommand),
+    Help,
+    Version,
+}
+
+#[cfg(test)]
+#[path = "main_tests.rs"]
+mod main_tests;
+
+impl Mode {
+    fn parse(args: Vec<String>) -> Result<Self> {
+        match args.as_slice() {
+            [] => Ok(Self::ServeMcp),
+            [flag] if flag == "--help" || flag == "-h" || flag == "help" => Ok(Self::Help),
+            [flag] if flag == "--version" || flag == "-V" || flag == "version" => Ok(Self::Version),
+            [command] if command == "mcp" => Ok(Self::StdioMcp),
+            [serve, service] if serve == "serve" && service == "mcp" => Ok(Self::ServeMcp),
+            [command, rest @ ..]
+                if matches!(
+                    command.as_str(),
+                    "search" | "tail" | "errors" | "hosts" | "correlate" | "stats"
+                ) =>
+            {
+                let mut cli_args = Vec::with_capacity(rest.len() + 1);
+                cli_args.push(command.clone());
+                cli_args.extend(rest.iter().cloned());
+                Ok(Self::Cli(cli::CliCommand::parse(cli_args)?))
+            }
+            _ => {
+                print_usage();
+                anyhow::bail!("unknown command: {}", args.join(" "));
+            }
+        }
+    }
+
+    fn default_log_filter(&self) -> &'static str {
+        match self {
+            Self::ServeMcp => "info",
+            Self::StdioMcp => "warn",
+            Self::Cli(_) => "warn",
+            Self::Help => "info",
+            Self::Version => "info",
+        }
+    }
+}
+
+fn print_usage() {
+    eprintln!(
+        "Usage:
+  hive --version     Print version
+  hive serve mcp     Start syslog UDP/TCP ingest plus HTTP MCP server
+  hive mcp           Start query-only MCP stdio transport
+  hive search [query] [--hostname HOST] [--source-ip SOURCE] [--severity LEVEL] [--app-name APP] [--from TIME] [--to TIME] [--limit N] [--json]
+  hive tail [-n N] [--hostname HOST] [--source-ip SOURCE] [--app-name APP] [--json]
+  hive errors [--from TIME] [--to TIME] [--json]
+  hive hosts [--json]
+  hive correlate --reference-time TIME [--window-minutes N] [--severity-min LEVEL] [--hostname HOST] [--source-ip SOURCE] [--query FTS] [--limit N] [--json]
+  hive stats [--json]
+
+Legacy alias:
+  syslog            Runs the same commands for this transition release
+
+Environment:
+  HIVE_MCP_DB_PATH     SQLite database path used by both transports
+  SYSLOG_MCP_DB_PATH   Legacy alias for HIVE_MCP_DB_PATH
+  RUST_LOG             Log filter; stdio logs always go to stderr"
+    );
+}
+
+async fn shutdown_signal() {
+    let ctrl_c = async {
+        match tokio::signal::ctrl_c().await {
+            Ok(()) => {}
+            Err(e) => {
+                tracing::error!(error = %e, "Failed to install CTRL+C handler");
+                std::future::pending::<()>().await;
+            }
+        }
+    };
+
+    #[cfg(unix)]
+    let terminate = async {
+        match tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate()) {
+            Ok(mut sig) => {
+                sig.recv().await;
+            }
+            Err(e) => {
+                tracing::error!(error = %e, "Failed to install SIGTERM handler");
+                std::future::pending::<()>().await;
+            }
+        }
+    };
+
+    #[cfg(not(unix))]
+    let terminate = std::future::pending::<()>();
+
+    tokio::select! {
+        _ = ctrl_c => {},
+        _ = terminate => {},
+    }
+    tracing::info!("Shutdown signal received");
+}
 
 /// Test support: factory helpers for building [`mcp::AppState`] variants.
 ///
